Extract Height type to primitives/height.rs with safety tests
Moves Height type and implementations from legacy.rs to dedicated module:
- Creates types/db/primitives/height.rs
- Adds 5 unit tests proving critical safety bugs in current implementation
- Updates primitives.rs to export Height and GENESIS_HEIGHT

Tests prove:
1. Underflow wraps to u32::MAX in release (Height(0) - 1 = Height(4294967295))
2. Underflow panics in debug mode
3. Overflow wraps in release (Height(u32::MAX-10) + 20 = Height(9))
4. Overflow panics in debug mode
5. Arithmetic bypasses TryFrom validation (can create Heights > MAX)

Real-world impact documented: non_finalised_state.rs:390 reorg detection bug
where unchecked subtraction causes silent failure instead of error.

This extraction prepares for future migration to safe arithmetic that returns
Option, matching Zebra's Height implementation.
